Move labeling code to myia.debug.label
